Use model dtype for timestep in load_inputs
HunyuanVideo15TimeEmbedding.forward casts the sinusoidal projection to
timestep.dtype before passing it to the linear layer, so the timestep
must have the same dtype as the model weights (bfloat16) to avoid a
dtype mismatch error.

+++ hunyuan_video_1_5_i2v_720p_gguf/pytorch/loader.py
@@ -163,10 +163,10 @@ class ModelLoader(ForgeModel):
             config.image_embed_dim,
             dtype=dtype,
         )
-        # Use float32 timestep: HunyuanVideo15TimeEmbedding.forward casts the
-        # sinusoidal projection to timestep.dtype, so Long would produce a Long
-        # tensor that cannot be multiplied by bfloat16 weights.
-        timestep = torch.tensor([500.0], dtype=torch.float32).expand(batch_size)
+        # Use bfloat16 timestep: HunyuanVideo15TimeEmbedding.forward casts the
+        # sinusoidal projection to timestep.dtype, so the timestep dtype must
+        # match the model weights dtype to avoid a dtype mismatch in linear_1.
+        timestep = torch.tensor([500.0], dtype=dtype).expand(batch_size)
 
         return {
             "hidden_states": hidden_states,
